Get read/write to compile without warnings. There is a lot of debug that will go away and some API changes coming up.

+++ com32/gplinclude/disk/read.h
@@ -13,6 +13,6 @@
 #include <disk/geom.h>
 
 int read_mbr(int, void *);
-int dev_read(int, void *, unsigned int, int);
-int read_sectors(struct driveinfo *, void *, const unsigned int, const int);
+int dev_read(int, void *, unsigned long int, int);
+int read_sectors(struct driveinfo *, void *, const unsigned long int, const int);
 #endif /* _READ_H */
